updated to object oriented

+++ api/oop-compiler.php
@@ -4,7 +4,7 @@
 class compiler{
     var $file_name;
     // define("tmp","./tmp/");
-    var $tmp="./tmp/";
+    var $tmp="./api/tmp/";
     var $compiled=false;
     function create_file($code)
     {
@@ -17,8 +17,9 @@ class compiler{
     }
     function compile()
     {
-        $compile_cmd = "cd tmp && g++ $this->file_name.cpp -o $this->file_name.exe 2> $this->file_name.err" ;
-        system($compile_cmd);
+        $compile_cmd = "cd $this->tmp && g++ $this->file_name.cpp -o $this->file_name.exe 2> $this->file_name.err" ;
+        // echo $compile_cmd;
+        $exec_res=exec($compile_cmd);
         if(filesize($this->tmp.$this->file_name.".err")>0)
         {
             $error_handle=fopen($this->tmp.$this->file_name.".err","r");
@@ -36,20 +37,19 @@ class compiler{
         $file_resource=fopen($this->tmp.$this->file_name.".input","w");
         fwrite($file_resource,$test_input);
         fclose($file_resource);
-        $execute_cmd = "cd tmp && start \"$this->file_name\" /B $this->file_name.exe < $this->file_name.input > $this->file_name.output & timeout /t 0.1 & Taskkill /im $this->file_name.exe /f" ;
+        $execute_cmd = "cd $this->tmp && start \"$this->file_name\" /B $this->file_name.exe < $this->file_name.input > $this->file_name.output & timeout /t 3 & Taskkill /im $this->file_name.exe /f" ;
         $exec_res=exec($execute_cmd);
-        unlink($this->tmp.$this->file_name.".input");
+        if(file_exists($this->tmp.$this->file_name.".input")) unlink($this->tmp.$this->file_name.".input"); 
         if(!$exec_res)
         {
             $output_handle=fopen($this->tmp.$this->file_name.".output","r");
             $output_file_size=filesize($this->tmp.$this->file_name.".output");
             if($output_file_size>0) $retval=fread($output_handle,$output_file_size);
-            else $retval="Run time Error";
+            else $retval="Run time Error or TLE";
             fclose($output_handle);
-            unlink($this->tmp.$this->file_name.".output");
-            return $retval;
+            $exec_res=$retval;
         }
-        unlink($this->tmp.$this->file_name.".output");
+        if(file_exists($this->tmp.$this->file_name.".output")) unlink($this->tmp.$this->file_name.".output");
         return $exec_res;
     }
     function close()
@@ -62,25 +62,25 @@ class compiler{
                     unlink($this->tmp.$this->file_name.".exe");
                 }
                 catch(Exception $e){
-                    sleep(1);
+                    //sleep(1);
                 }
             }
             // while(unlink($this->tmp.$this->file_name.".exe"));
         }
-        unlink($this->tmp.$this->file_name.".cpp");
+        if(file_exists($this->tmp.$this->file_name.".cpp")) unlink($this->tmp.$this->file_name.".cpp");
     }
 }
 
 
 // $t=new compiler();
-// // $code="#include<stdio.h>
+// $code="#include<stdio.h>
 
-// // int main()
-// // {
-// //     int *p=0;
-// //     printf(\"%d\",*p);
-// //     return 0;
-// // }";
+// int main()
+// {
+//     int *p=0;
+//     printf(\"%d\",5);
+//     return 0;
+// }";
 
 // $t->create_file($code);
 // if(($res=$t->compile())=="Success"){
